refactor selectors
