+++ apollo-router/src/plugins/telemetry/config_new/extendable.rs
@@ -287,11 +287,11 @@ mod test {
     use crate::plugins::telemetry::config_new::http_common::attributes::HttpCommonAttributes;
     use crate::plugins::telemetry::config_new::http_server::attributes::HttpServerAttributes;
     use crate::plugins::telemetry::config_new::router::attributes::RouterAttributes;
+    use crate::plugins::telemetry::config_new::router::selectors::RouterSelector;
     use crate::plugins::telemetry::config_new::selectors::OperationName;
     use crate::plugins::telemetry::config_new::selectors::ResponseStatus;
-    use crate::plugins::telemetry::config_new::selectors::RouterSelector;
-    use crate::plugins::telemetry::config_new::selectors::SupergraphSelector;
     use crate::plugins::telemetry::config_new::supergraph::attributes::SupergraphAttributes;
+    use crate::plugins::telemetry::config_new::supergraph::selectors::SupergraphSelector;
 
     #[test]
     fn test_extendable_serde() {
